packages/order_food_server/requirements.txt
Django>=5.0.6,<6.0
djangorestframework>=3.12,<4.0
djangorestframework-simplejwt>=4.7,<5.0
drf-yasg>=1.20,<2.0
django-cors-headers>=3.7,<4.0
psycopg2-binary>=2.8,<3.0
django-rest-swagger>=2.2,<3.0
django-sslserver>=0.22,<1.0
python-decouple==3.8
google-api-core==2.19.1
google-api-python-client==2.135.0
google-auth==2.25.2
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
gspread==6.1.2
utils==1.0.1
virtualenv==20.26.3
payos==0.1.5
